+++ packages/pinning-aggregation/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-aggregation@0.1.3...@ceramicnetwork/pinning-aggregation@0.1.4) (2020-12-08)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-aggregation
+
+
+
+
+
 ## [0.1.3](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-aggregation@0.1.2...@ceramicnetwork/pinning-aggregation@0.1.3) (2020-12-01)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-aggregation